+++ RELEASE_NOTES.md
@@ -2,11 +2,17 @@
 
 ## Summary
 
-<!-- Here goes a general summary of what this release is about -->
+The project has a new home!
+
+https://frequenz-floss.github.io/frequenz-channels-python/
+
+For now the documentation is pretty scarce but we will be improving it with
+time.
 
 ## Upgrading
 
-<!-- Here goes notes on how to upgrade from previous versions, including if there are any depractions and what they should be replaced with --> 
+* You need to make sure to use [timezone-aware] `datetime` objects when using
+  the timestamp returned by [`Timer`], Otherwise you will get an exception.
 
 ## New Features
 
@@ -14,4 +20,9 @@
 
 ## Bug Fixes
 
-<!-- Here goes notable bug fixes that are worth a special mention or explanation -->
+* [`Timer`] now returns [timezone-aware] `datetime` objects using UTC as
+  timezone.
+
+
+[`Timer`]: https://frequenz-floss.github.io/frequenz-channels-python/v0.11/reference/frequenz/channels/#frequenz.channels.Timer
+[timezone-aware]: https://docs.python.org/3/library/datetime.html#aware-and-naive-objects